+++ Sources/MapboxNavigation/NavigationViewController.swift
@@ -33,6 +33,7 @@ public enum MapOrnamentPosition {
  `CarPlayNavigationViewController` manages the corresponding user interface on a CarPlay screen.
  */
 open class NavigationViewController: UIViewController, NavigationStatusPresenter {
+    
     /**
      A `Route` object constructed by [MapboxDirections](https://docs.mapbox.com/ios/api/directions/) along with its index in a `RouteResponse`.
       
@@ -50,6 +51,8 @@ open class NavigationViewController: UIViewController, NavigationStatusPresenter
             }
         }
     }
+    
+    var didChangeAuthorizationIsFirstCalled = true
     
     /**
      A `Route` object constructed by [MapboxDirections](https://docs.mapbox.com/ios/api/directions/).
@@ -401,7 +404,10 @@ open class NavigationViewController: UIViewController, NavigationStatusPresenter
         guard AVAudioSession.sharedInstance().outputVolume <= NavigationViewMinimumVolumeForWarning else { return }
         
         let title = NSLocalizedString("INAUDIBLE_INSTRUCTIONS_CTA", bundle: .mapboxNavigation, value: "Adjust Volume to Hear Instructions", comment: "Label indicating the device volume is too low to hear spoken instructions and needs to be manually increased")
-        showStatus(title: title, spinner: false, duration: 3, animated: true, interactive: false)
+        
+        // create low volume notification status and append to array of statuses
+        let lowVolumeStatus = StatusView.Status(identifier: "INAUDIBLE_INSTRUCTIONS_CTA", title: title, duration: 3, animated: true, priority: 3)
+        show(lowVolumeStatus)
     }
     
     
@@ -462,9 +468,28 @@ open class NavigationViewController: UIViewController, NavigationStatusPresenter
         UNUserNotificationCenter.current().add(notificationRequest, withCompletionHandler: nil)
     }
     
-    public func showStatus(title: String, spinner: Bool, duration: TimeInterval, animated: Bool, interactive: Bool) {
+    /**
+     Shows a Status for a specified amount of time.
+     */
+    public func show(_ status: StatusView.Status) {
         navigationComponents.compactMap({ $0 as? NavigationStatusPresenter }).forEach {
-            $0.showStatus(title: title, spinner: spinner, duration: duration, animated: animated, interactive: interactive)
+            $0.show(status)
+        }
+    }
+    
+    /**
+     Hides a given Status without hiding the status view.
+     */
+    public func hide(_ status: StatusView.Status) {
+        navigationComponents.compactMap({ $0 as? NavigationStatusPresenter }).forEach {
+            $0.hide(status)
+        }
+    }
+    
+    @available(*, deprecated, message: "Add a status using show(_:) instead")
+    public func showStatus(title: String, spinner spin: Bool, duration: TimeInterval, animated: Bool, interactive: Bool) {
+        navigationComponents.compactMap({ $0 as? NavigationStatusPresenter }).forEach {
+            $0.showStatus(title: title, spinner: spin, duration: duration, animated: animated, interactive: interactive)
         }
     }
 }
@@ -724,21 +749,29 @@ extension NavigationViewController: NavigationServiceDelegate {
     public func navigationServiceShouldDisableBatteryMonitoring(_ service: NavigationService) -> Bool {
         return navigationComponents.allSatisfy { $0.navigationServiceShouldDisableBatteryMonitoring(service) }
     }
-    
+        
     public func navigationServiceDidChangeAuthorization(_ service: NavigationService, didChangeAuthorizationFor locationManager: CLLocationManager) {
         // CLLocationManager.accuracyAuthorization was introduced in the iOS 14 SDK in Xcode 12, so Xcode 11 doesn’t recognize it.
         guard let accuracyAuthorizationValue = locationManager.value(forKey: "accuracyAuthorization") as? Int else { return }
         let accuracyAuthorization = MBNavigationAccuracyAuthorization(rawValue: accuracyAuthorizationValue)
+        let previousAuthorizationValue = 1 - accuracyAuthorizationValue
+                        
+        // create authorization status
+        let title = NSLocalizedString("ENABLE_PRECISE_LOCATION", bundle: .mapboxNavigation, value: "Enable precise location to navigate", comment: "Label indicating precise location is off and needs to be turned on to navigate")
+        let authorizationStatus = StatusView.Status(identifier: "ENABLE_PRECISE_LOCATION", title: title, duration: .infinity, priority: 1)
         
         if #available(iOS 14.0, *), accuracyAuthorization == .reducedAccuracy {
-            let title = NSLocalizedString("ENABLE_PRECISE_LOCATION", bundle: .mapboxNavigation, value: "Enable precise location to navigate", comment: "Label indicating precise location is off and needs to be turned on to navigate")
-            showStatus(title: title, spinner: false, duration: 20, animated: true, interactive: false)
+            show(authorizationStatus)
             mapView?.reducedAccuracyActivatedMode = true
+        } else if #available(iOS 14.0, *), previousAuthorizationValue == 1, didChangeAuthorizationIsFirstCalled == false {
+            hide(authorizationStatus)
+            mapView?.reducedAccuracyActivatedMode = false
         } else {
             //Fallback on earlier versions
             mapView?.reducedAccuracyActivatedMode = false
             return
         }
+        didChangeAuthorizationIsFirstCalled = false
     }
     
     // MARK: - Building Extrusion Highlighting